test order state tracking

+++ app/code/local/Orba/Shipping/Helper/Carrier/Tracking.php
@@ -204,8 +204,8 @@ class Orba_Shipping_Helper_Carrier_Tracking extends Mage_Core_Helper_Abstract {
                 Mage::log('Po id: ' . $orderPo->getId(), null, 'tracking.log');
                 Mage::log('Status: ' . $orderPo->getUdropshipStatus(), null, 'tracking.log');
                 Mage::log('Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED: ' . Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED, null, 'tracking.log');
-                Mage::log('Eq: ' . (int)($orderPo->getUdropshipStatus() !== Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED), null, 'tracking.log');
-                if($orderPo->getUdropshipStatus() !== Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED){
+                Mage::log('Eq: ' . ((int)$orderPo->getUdropshipStatus() !== (int)Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED), null, 'tracking.log');
+                if((int)$orderPo->getUdropshipStatus() !== (int)Unirgy_Dropship_Model_Source::SHIPMENT_STATUS_DELIVERED){
                     $orderCompleted = false;
                 }
             }
